# project2/services/notification/src/main/java/com/insurance/notification/model/ErrorResponse.java
package com.insurance.notification.model;

import java.time.LocalDateTime;
import java.util.List;

public record ErrorResponse(
        int status,
        String error,
        String message,
        List<String> details,
        LocalDateTime timestamp
) {
    public ErrorResponse(int status, String error, String message) {
        this(status, error, message, List.of(), LocalDateTime.now());
    }

    public ErrorResponse(int status, String error, String message, List<String> details) {
        this(status, error, message, details, LocalDateTime.now());
    }
}